+++ app/Providers/AppServiceProvider.php
@@ -2,9 +2,11 @@
 
 namespace App\Providers;
 
+use App\Contracts\ActivityLoggerInterface;
 use App\Contracts\LicenseServiceInterface;
 use App\Models\User;
 use App\Observers\UserObserver;
+use App\Services\ActivityLogger;
 use App\Services\LicenseService;
 use App\Translation\DatabaseLoader;
 use Illuminate\Support\Facades\Schema;
@@ -20,6 +22,7 @@ class AppServiceProvider extends ServiceProvider
         });
 
         $this->app->bind(LicenseServiceInterface::class, LicenseService::class);
+        $this->app->bind(ActivityLoggerInterface::class, ActivityLogger::class);
     }
 
     /**